refactor API endpoints to use relative paths
