+++ gitarena-frontend/lib/instance-config.ts
@@ -1,5 +1,3 @@
-import { cache } from "react";
-
 export interface InstanceConfig {
     app: string;
     version: string;
@@ -8,15 +6,3 @@ export interface InstanceConfig {
     repository: string;
     commit: string;
 }
-
-export const getInstanceConfig = cache(async (): Promise<InstanceConfig | null> => {
-    try {
-        const res = await fetch("http://localhost:8080/api", { cache: "force-cache" });
-        if (!res.ok) {
-            return null;
-        }
-        return res.json();
-    } catch {
-        return null;
-    }
-});
